refactor(grove): rename EnvSyncDialog to ProjectSettingsDialog
- Reposition env sync as a section within Project Settings dialog
- Add cursor-pointer to dialog X button and action buttons
- Fix dialog repositioning by adding min-height
- Polish UI: skeleton loading, hover states, font-mono patterns,
  visual hierarchy with section heading and description
